+++ shared/android-core/src/main/kotlin/AndroidSdk.kt
@@ -1,7 +1,0 @@
-package com.javiersc.plugins.android.core
-
-object AndroidSdk {
-
-    const val compileSdk = 31
-    const val minSdk = 21
-}