+++ shared/core/src/main/kotlin/GroupId.kt
@@ -1,6 +1,0 @@
-package com.javiersc.plugins.core
-
-import org.gradle.api.Project
-
-val Project.groupId: String
-    get() = "${property("allProjects.group")}.${property("allProjects.name")}"