+++ shared/core/src/main/kotlin/IsSignificantVersion.kt
@@ -1,4 +1,6 @@
-package com.javiersc.plugins.core
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.javiersc.gradle.plugins.core
 
 import org.gradle.api.Project
 
@@ -9,3 +11,6 @@ val Project.isSignificant: Boolean
                     "[0-9a-zA-Z-]*)(?:\\.(?:[1-9]\\d*|\\d*[a-zA-Z-][0-9]*))*))?(?:\\+([0-9]))?\$",
             )
             .matches(version.toString())
+
+val Project.onlySignificantGradleProperty: Boolean
+    get() = property("publish.onlySignificant")?.toString()?.toBoolean() ?: true